requirements.txt
flask
requests
gunicorn
pytest
celery
redis
prometheus_client
python-dotenv
ruff
mypy
